for #7332 moved radia specific code to radia.js and removed unused code
